4.21: blogilistan laajennus, step 10

+++ controllers/blogs.js
@@ -61,9 +61,25 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+
   try{
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!request.token || !decodedToken.id){
+      return response.status(401).json({Error: 'Token missing or invalid'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+    
+    if(blog.user.toString() === user._id.toString()){
+      await Blog.findByIdAndRemove(request.params.id)
+      user.blogs.splice(user.blogs.indexOf(request.params.id), 1)
+      await user.save()
+      response.status(204).end()
+    }else{
+      response.status(401).json({Error: 'Token missing or invalid xxx'})
+    }
   }catch(exception){
     next(exception)
   }
